+++ Changes.md
@@ -410,7 +410,10 @@ Build
 - USD : Updated to version 23.05.
 - ZLib : Added version 1.2.13.
 
-1.2.10.x (relative to 1.2.10.3)
+1.2.10.x (relative to 1.2.10.4)
+========
+
+1.2.10.4 (relative to 1.2.10.3)
 ========
 
 Fixes